Move storagespace model into own file
