+++ archive/common/src/main/scala/uk/ac/wellcome/platform/archive/common/models/BagId.scala
@@ -1,8 +1,8 @@
 package uk.ac.wellcome.platform.archive.common.models
 
 case class BagId(
-  space: StorageSpace,
-  externalIdentifier: ExternalIdentifier
+                  space: StorageSpace,
+                  externalIdentifier: ExternalIdentifier
 ) {
   override def toString: String =
     s"$space/$externalIdentifier"
